fix: split structural test-evidence tests out to fix debt-no-growth gate
CI's architecture step failed: tests/test_bugfix_test_contract.py is
tracked in architecture/debt.yaml's ADR-061 no-growth ledger at a
1267-line baseline (a pre-existing legacy_oversized_test_module entry),
and several rounds of review findings on the structural test-evidence
classification logic (standalone CI-invoked runners, a non-gating
script, conftest.py at any depth) pushed it to 1361 lines -- 94 over.
The gate's own guidance: move responsibility instead of raising the
baseline.

Moved TestStructuralRequirement (touches_shipped_code/touches_tests/
is_test_path/adds_or_modifies_a_test and the file-shape predicates
underneath them) verbatim into a new sibling file,
tests/test_bugfix_test_contract_evidence.py -- not itself tracked in
the debt ledger, so it carries no baseline of its own. The original
file is now 1206 lines, back under its 1267 baseline with room to
spare; the same 265 tests pass, just split across two files.

Verified against CI's own base sha (327df7b, the PR base):
ARCHITECTURE_BASE=327df7b5616bcfaea8c330aad418b796c17f3970 python
scripts/check_architecture.py now reports 0 errors.
